Move stryker4s.run.report package to stryker4s.report
