+++ core/src/test/scala/stryker4s/report/ConsoleReporterTest.scala
@@ -1,4 +1,4 @@
-package stryker4s.run.report
+package stryker4s.report
 import java.io.File
 import java.nio.file.Paths
